refactor: pull ImporterFileType into separate class
Moved the `ImporterFileType` enum from `ImporterDirectoryWatcher` to a standalone class for better modularity and reusability. Updated all references to point to the new location. This change simplifies code dependencies and improves maintainability.
