+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/coder/stream/Asn1CodecPublisher.java
@@ -19,7 +19,7 @@ package us.dot.its.jpo.ode.coder.stream;
 import java.io.BufferedInputStream;
 import java.util.List;
 import us.dot.its.jpo.ode.coder.stream.LogFileToAsn1CodecPublisher.LogFileToAsn1CodecPublisherException;
-import us.dot.its.jpo.ode.importer.ImporterDirectoryWatcher.ImporterFileType;
+import us.dot.its.jpo.ode.importer.ImporterFileType;
 import us.dot.its.jpo.ode.importer.parser.LogFileParser;
 import us.dot.its.jpo.ode.importer.parser.LogFileParserFactory;
 import us.dot.its.jpo.ode.model.OdeData;
