+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/coder/stream/DecoderPublisher.java
@@ -16,8 +16,7 @@
 package us.dot.its.jpo.ode.coder.stream;
 
 import java.io.BufferedInputStream;
-
-import us.dot.its.jpo.ode.importer.ImporterDirectoryWatcher.ImporterFileType;
+import us.dot.its.jpo.ode.importer.ImporterFileType;
 
 public interface DecoderPublisher {
