+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/importer/ImporterFileType.java
@@ -1,0 +1,5 @@
+package us.dot.its.jpo.ode.importer;
+
+public enum ImporterFileType {
+  LOG_FILE, JSON_FILE
+}
